Allow tasklink to be able to handle Assessments
As this function mainly generates markup I think it makes sense to
reuse it as opposed to duplicating it.
