+++ server/views/assessments/show.njk
@@ -28,7 +28,7 @@
               {% for task in section.tasks %}
                 <li class="app-task-list__item">
                   <span class="app-task-list__task-name">
-                    {{ taskLink(task, assessment)  }}
+                    {{ taskLink(task, assessment, 'assessments')  }}
                   </span>
                   {{ getTaskStatus(task, assessment) }}
                 </li>
